Added Kick player.
Currently a script checks if stream is live every 10 seconds so far it seems there's no rate limit on their API but this is not a good solution.
Added a button on top to switch to Kick, only appears if Kick stream is detected as 'live'.
Select channel on page load, defaults to XQC.